+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # vuex-loading changelog
 
-## UNRELEASED
+## v1.0.0
+
+- Remove built-in loaders
+- Remove `width` and `height` props.
+- Strict props.
+- Loader supports RegExp
+- Rename `registerComponents` to `registerComponent`
+- added `helperName` option
+
+## v0.4.0
 
 - rename v-loading slot `spinner` to `loading` #30
 - added `wrapLoading` helper function for easy integration of vuex-loading in vue component methods #30